+++ lib/graphql_flutter.dart
@@ -1,5 +1,9 @@
 library graphql_flutter;
 
 export './src/client.dart';
-export './src/query.dart';
-export './src/mutation.dart';
+
+export './src/cache/in_memory.dart';
+
+export './src/widgets/query.dart';
+export './src/widgets/mutation.dart';
+export './src/widgets/cache_provider.dart';